Move tests to test suite crate
Closes #68.

Also, removed travis-cargo now that docs are handled by docs.rs and
don't need to be uploaded from travis.

Minor update to the compiletest script to avoid issues with stale
artifacts left around after `rustup update`.
