+++ .travis.yml
@@ -2,6 +2,7 @@ os:
   - linux
   - osx
 language: rust
+cache: cargo
 rust:
   # This version is tested to avoid unintentional bumping of the minimum supported Rust version
   - 1.20.0
@@ -9,21 +10,12 @@ rust:
   - beta
   - nightly
 sudo: false
-before_script:
-  - pip install -v 'travis-cargo<0.2' --user && export PATH=$HOME/.local/bin:$PATH
-  - if [[ -e ~/Library/Python/2.7/bin ]]; then export PATH=~/Library/Python/2.7/bin:$PATH; fi
 script:
-  - travis-cargo build
-  - travis-cargo test
-  - travis-cargo --only nightly test -- --all
-  - travis-cargo --only stable doc
-after_success:
-  - travis-cargo --only nightly doc-upload
+  - cargo test
+  - if [ "$TRAVIS_RUST_VERSION" = nightly ]; then (cd ./test_suite && cargo test --features unstable); fi
 env:
   global:
-    - TRAVIS_CARGO_NIGHTLY_FEATURE=unstable_testing
     - secure: "DoZ8g8iPs+X3xEEucke0Ae02JbkQ1qd1SSv/L2aQqxULmREtRcbzRauhiT+ToQO5Ft1Lul8uck14nPfs4gMr/O3jFFBhEBVpSlbkJx7eNL3kwUdp95UNroA8I43xPN/nccJaHDN6TMTD3+uajTQTje2SyzOQP+1gvdKg17kguvE="
-
 
 
 notifications:
