--- compiletest/Cargo.toml
@@ -1,7 +1,0 @@
-[project]
-name = "bitflags-compiletest"
-version = "0.0.0"
-
-[dev-dependencies]
-bitflags = { path = "../" }
-compiletest_rs = { version = "0.2" }
