+++ test_suite/tests/compile-fail/private_flags.rs
