+++ test_suite/tests/external_no_std.rs
