+++ test_suite/tests/i128_bitflags.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "unstable_testing")]
+#![cfg(feature = "unstable")]
 
 #![feature(i128_type)]
